Create monotor
