+++ internal/handler/static/middleware_options.go
@@ -2,7 +2,7 @@ package static
 
 import (
 	"github.com/evg4b/uncors/internal/contracts"
-	"github.com/evg4b/uncors/internal/tui/request_tracker"
+	"github.com/evg4b/uncors/internal/tui/monitor"
 	"github.com/spf13/afero"
 )
 
@@ -32,7 +32,7 @@ func WithPrefix(prefix string) MiddlewareOption {
 	}
 }
 
-func WithRequestTracker(tracker request_tracker.RequestTracker) MiddlewareOption {
+func WithRequestTracker(tracker monitor.RequestTracker) MiddlewareOption {
 	return func(h *Middleware) {
 		h.tracker = tracker
 	}
