+++ internal/tui/monitor/responces.go
@@ -1,4 +1,4 @@
-package request_tracker
+package monitor
 
 import (
 	"sort"
