+++ internal/uncors/app.go
@@ -10,7 +10,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/evg4b/uncors/internal/tui/request_tracker"
+	"github.com/evg4b/uncors/internal/tui/monitor"
 
 	"github.com/charmbracelet/log"
 	"github.com/evg4b/uncors/internal/config"
@@ -33,7 +33,7 @@ type App struct {
 	httpsListenerMutex *sync.Mutex
 	httpsListener      net.Listener
 	cache              appCache
-	tracker            request_tracker.RequestTracker
+	tracker            monitor.RequestTracker
 }
 
 const (
@@ -42,7 +42,7 @@ const (
 	shutdownTimeout   = 15 * time.Second
 )
 
-func CreateApp(fs afero.Fs, version string, tracker request_tracker.RequestTracker) *App {
+func CreateApp(fs afero.Fs, version string, tracker monitor.RequestTracker) *App {
 	return &App{
 		fs:                 fs,
 		version:            version,
