+++ internal/uncors/keymap.go
@@ -1,0 +1,32 @@
+package uncors
+
+import "github.com/charmbracelet/bubbles/key"
+
+// keyMap defines a set of keybindings. To work for help it must satisfy
+// key.Map. It could also very easily be a map[string]key.Binding.
+type keyMap struct {
+	Quit    key.Binding
+	Restart key.Binding
+}
+
+// ShortHelp returns keybindings to be shown in the mini help view. It's part
+// of the key.Map interface.
+func (k keyMap) ShortHelp() []key.Binding {
+	return []key.Binding{
+		k.Quit,
+		k.Restart,
+	}
+}
+
+// FullHelp returns keybindings for the expanded help view. It's part of the
+// key.Map interface.
+func (k keyMap) FullHelp() [][]key.Binding {
+	return [][]key.Binding{
+		{k.Quit, k.Restart}, // second column
+	}
+}
+
+var keys = keyMap{
+	Restart: key.NewBinding(key.WithKeys("r"), key.WithHelp("r", "restart server")),
+	Quit:    key.NewBinding(key.WithKeys("q", "ctrl+c"), key.WithHelp("q", "quit")),
+}
